remove addon: SABnzbd-Suite
ours is broken and not maintained.
I recommend using lsellens' addon. and well most people already
use that, anyway. so let's not conflict with his stuff and not
confuse users..

if requested, this can be re-added (git never forgets)

dependent python packages, if re-added, should be converted to
xbmc.python.script addons. I can and will help with that if someone
needs them.